requirements.txt
flask>=3.0
requests>=2.31
python-dotenv>=1.0
yfinance>=0.2.36
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
